Add files via upload
Attached is the picture of a case of imperfection in the structure (data for 10 different cases was available and merged) [data not uploaded due to privacy].
RCF and FFNN was tried and tested and subsequent imperfect nodes were determined. Additionally the report and presentation for the project is attached for high level overview.